Tratativas de erros

+++ controllers/PetController.js
@@ -3,7 +3,7 @@ const Owner = require('../models/Owner')
 
 class PetController {
 
-    static async createPet(req, res){
+    static async createPet(req, res) {
         const pet = {
             name: req.body.name,
             breed: req.body.breed,
@@ -12,42 +12,85 @@ class PetController {
             notes: req.body.notes,
             OwnerId: req.body.OwnerId
         }
+
+        if (pet.name === "" || pet.breed === "" || pet.age === "" || pet.color === "" || pet.notes === "" || pet.OwnerId === "") {
+            res.status(402).json({ message: 'pet-parameter-null' })
+            return;
+        };
+
+        const owner = await Owner.findOne({ where: { id: pet.OwnerId } })
+        if (!owner) {
+            res.status(401).json({ message: `owner-${pet.OwnerId}-no registry` })
+            return;
+        }
+
         await Pet.create(pet)
-        res.status(202).json({ message: `pet-${pet.name}-created`})
+        res.status(202).json({ message: `pet-${pet.name}-created` })
     }
 
-    static async showPets(req, res){
-        const owner = await Owner.findOne({include: Pet, where: { id: req.params.id }})
-    
+    static async showPets(req, res) {
+
+        const owner = await Owner.findOne({ include: Pet, where: { id: req.params.id } })
+
+        if (!owner) {
+            res.status(401).json({ message: 'pet-owner-invalid' })
+            return;
+        };
+
+
         res.status(201).json({ owner })
         // ({ owner: owner.get({ plain: true })})
-    
+
     }
 
 
-    static async listPetToUpdate(req, res){
-        const pet = await Pet.findOne({ where: { id: req.params.id }})
+    static async listPetToUpdate(req, res) {
+        const pet = await Pet.findOne({ where: { id: req.params.id } })
+
+        if (!pet) {
+            res.status(402).json({ message: 'pet-parameter-null' })
+            return;
+        }
 
         res.status(200).json(pet)
     }
 
     static async updatePet(req, res) {
-    
+
         const pet = {
-            name: req.body.name,        
+            name: req.body.name,
             breed: req.body.breed,
             age: req.body.age,
             color: req.body.color,
             notes: req.body.notes,
             OwnerId: req.body.OwnerId
         }
-        await Pet.update(pet, { where: { id: req.body.id }})
+
+        if (pet.name === "" || pet.breed === "" || pet.age === "" || pet.color === "" || pet.notes === "" || pet.OwnerId === "") {
+            res.status(402).json({ message: 'pet-parameter-null' })
+            return;
+        };
+
+        const owner = await Owner.findOne({ where: { id: pet.OwnerId } })
+        if (!owner) {
+            res.status(401).json({ message: `owner-${pet.OwnerId}-parameter-null` })
+            return;
+        }
+
+        await Pet.update(pet, { where: { id: req.body.id } })
         res.status(200).json({ message: `pet-${pet.name}-successfully-updated` })
     }
 
     static async deletePet(req, res) {
 
-        await Pet.destroy({ where: { id: req.body.id }})
+        const pet = await Pet.findOne({ where: { id: req.body.id } })
+
+        if (!pet) {
+            res.status(401).json({ message: `pet-parameter-null` })
+            return;
+        }
+
+        await Pet.destroy({ where: { id: req.body.id } })
         res.status(202).json({ message: `pet-${req.body.id}-deleted` })
     }
 }
